feat(Quiz): create reusable `QuizManager`
