+++ src/components/Quiz/QuizzesList/index.tsx
@@ -7,7 +7,7 @@ import { trackCustomEvent } from "@/lib/utils/matomo"
 
 import allQuizzesData from "@/data/quizzes"
 
-import Translation from "../Translation"
+import Translation from "../../Translation"
 
 import QuizItem from "./QuizItem"
